fix: Functional connection to db that randomizes 10 quiz questions for a selected quiz, with exception handling.

+++ week3/d5-mvc-jsp/d5_mvc_jsp_quiz/src/main/java/com/example/d5_mvc_jsp_quiz/repository/quiz/QuizResultRepositoryRowMapper.java
@@ -6,14 +6,14 @@ import org.springframework.stereotype.Component;
 
 @Component
 public class QuizResultRepositoryRowMapper implements RowMapper<QuizResult> {
-    @Override
-    public QuizResult mapRow(java.sql.ResultSet resultSet, int i) throws java.sql.SQLException {
-        return new QuizResult(
-                resultSet.getInt("id"),
-                resultSet.getInt("user_id"),
-                resultSet.getInt("quiz_id"),
-                resultSet.getString("start_time"),
-                resultSet.getString("end_time")
-        );
-    }
+  @Override
+  public QuizResult mapRow(java.sql.ResultSet resultSet, int i) throws java.sql.SQLException {
+    return new QuizResult(
+      resultSet.getLong("id"),
+      resultSet.getLong("user_id"),
+      resultSet.getLong("quiz_id"),
+      resultSet.getString("start_time"),
+      resultSet.getString("end_time")
+    );
+  }
 }
